requirements.txt
python-socketio===5.11.1
